pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hunt"
version = "0.1.0"
description = "AI-powered resume tailoring for job applications"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.3.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "rich>=13.7.0",
    "questionary>=2.0.1",
]

[project.scripts]
hunt = "hunt.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hunt*"]
